refactor: :recycle: use current db methods & genericise utilities
